refactor: Clean up scripting
