+++ docs/README.md
@@ -334,16 +334,12 @@ The following npm scripts are supported in this directory:
 | Script | Description |
 |--------|-------------|
 | `build` | Build the site; outputs to `public/` by default. |
-| `build:api` | `npm run build:api-rollup && npm run build:api-documentation` |
-| `build:api-documentation` | Convert package API reports (`.api.json` files) into Markdown. |
-| `build:api-documentation:multi-version` | `node ./api-markdown-documenter/index.js true` |
-| `build:api-rollup` | Runs `rollup-api-json.js` to produce rolled-up API data. See the script for more details. |
-| `build:api-rollup:multi-version` | Runs `node ./rollup-api-json-helper.js true` while passing true to the optional parameter to produce rolled-up API data for all versions specified in data/versions.json. |
-| `build:api:multi-version` | `npm run build:api-rollup:multi-version && npm run build:api-documentation:multi-version` |
+| `build:api-documentation` | `node ./api-markdown-documenter/index.js` |
 | `build:md-magic` | Updates generated content in Markdown files. |
 | `build:md-magic:code` | `node markdown-magic-code.js` |
 | `build:md-magic:website` | `node markdown-magic-website.js` |
 | `build:multi-version` | Build the site; outputs to `public/` by default. Includes API content from all major release branches. |
+| `build:redirects` | `node ./build-redirects.js` |
 | `build:repo-docs` | `npm run build:md-magic:code` |
 | `build:website` | `npm run build:api-rollup && npm run build:md-magic:website && npm run build:api-documentation && npm run hugo` |
 | `ci:build` | `npm run download && npm run build` |
