+++ docs/api-markdown-documenter/index.js
@@ -13,7 +13,6 @@
 const chalk = require("chalk");
 const path = require("path");
 const versions = require("../data/versions.json");
-const { buildRedirects } = require("./build-redirects");
 const { renderApiDocumentation } = require("./render-api-documentation");
 
 const renderMultiVersion = process.argv[2];
@@ -66,5 +65,3 @@ Promise.all(apiDocRenders).then(
 		process.exit(1);
 	},
 );
-
-buildRedirects();
