+++ docs/package.json
@@ -12,19 +12,14 @@
 	"license": "MIT",
 	"author": "Microsoft and contributors",
 	"scripts": {
-		"build": "npm run build:api && npm run build:md-magic && npm run hugo",
-		"build:api": "npm run build:api-rollup && npm run build:api-documentation",
+		"build": "concurrently npm:build:api npm:build:md-magic && npm run hugo",
 		"build:api-documentation": "node ./api-markdown-documenter/index.js",
-		"build:api-documentation:multi-version": "node ./api-markdown-documenter/index.js true",
-		"build:api-rollup": "node ./rollup-api-json-helper.js",
-		"build:api-rollup:multi-version": "node ./rollup-api-json-helper.js true",
-		"build:api:multi-version": "npm run build:api-rollup:multi-version && npm run build:api-documentation:multi-version",
 		"build:md-magic": "concurrently npm:build:md-magic:code npm:build:md-magic:website",
 		"build:md-magic:code": "markdown-magic --files \"**/*.md\" !docs --workingDirectory ..",
 		"build:md-magic:website": "markdown-magic",
-		"build:multi-version": "npm run build:api:multi-version && npm run build:md-magic && npm run hugo",
+		"build:redirects": "node ./build-redirects.js",
 		"build:repo-docs": "npm run build:md-magic:code",
-		"build:website": "npm run build:api-rollup && npm run build:md-magic:website && npm run build:api-documentation && npm run hugo",
+		"build:website": "concurrently npm:build:md-magic:website npm:build:api-documentation npm:build:redirects && npm run hugo",
 		"ci:build": "npm run download && npm run build",
 		"ci:linkcheck": "start-server-and-test ci:start http://localhost:1313 linkcheck:full",
 		"ci:start": "http-server ./public --port 1313 --silent",
